Refactor to reduce coupling between variants and components used
