+++ src/api.ts
@@ -145,6 +145,12 @@ interface SEOData {
     webTitle: string;
 }
 
+export enum Fronts {
+    Opinion = "email/opinion",
+    FilmToday = "email/film-today",
+    MediaBriefing = "email/media-briefing"
+}
+
 export interface Front {
     id: string;
     collections: Collection[];
