+++ src/components/tests/opinion/Opinion.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+import { Collection as ICollection } from "../../../api";
+import { VariantB } from "./VariantB";
+import { VariantC } from "./VariantC";
+
+export const Opinion: React.FC<{
+    frontId: string;
+    collections: ICollection[];
+    salt: string;
+    variant?: string;
+}> = ({ frontId, collections, salt, variant }) => {
+    if (variant === "c") {
+        return (
+            <VariantC
+                frontId={frontId}
+                collections={collections}
+                salt={salt}
+            ></VariantC>
+        );
+    }
+
+    return (
+        <VariantB
+            frontId={frontId}
+            collections={collections}
+            salt={salt}
+        ></VariantB>
+    );
+};
