+++ src/components/tests/opinion/collections/commentB/Grid.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import { ContributorImageWrapper, getContributor } from "./Card";
 import { palette } from "@guardian/src-foundations";
-import { TableRow, TableRowCell } from "../../../layout/Table";
-import { Padding } from "../../../layout/Padding";
-import { Content } from "../../../api";
-import { GridRow, partition } from "../../../layout/Grid";
-import { TableCSS } from "../../../css";
+import { TableRowCell, TableRow } from "../../../../../layout/Table";
+import { Padding } from "../../../../../layout/Padding";
+import { Content } from "../../../../../api";
+import { GridRow, partition } from "../../../../../layout/Grid";
 import { Card as CommentCard } from "./Card";
 import { LinkCard } from "./LinkCard";
 
@@ -14,11 +13,6 @@ interface CommentGridProps {
     salt: string;
     shouldShowGridImages: boolean;
 }
-
-const tableStyle: TableCSS = {
-    borderCollapse: "collapse",
-    borderSpacing: 0
-};
 
 export const Grid: React.FC<CommentGridProps> = ({
     content,
@@ -65,49 +59,20 @@ export const Grid: React.FC<CommentGridProps> = ({
                             shouldShowImage={shouldShowGridImages}
                         />
                     }
-                    leftStyles={{
-                        backgroundColor: palette.neutral[100],
-                        borderLeft: `1px solid ${palette.opinion.main}`
-                    }}
-                    rightStyles={{
-                        backgroundColor: palette.neutral[100],
-                        borderLeft: `1px solid ${palette.opinion.main}`
-                    }}
+                    leftStyles={{ backgroundColor: palette.opinion.faded }}
+                    rightStyles={{ backgroundColor: palette.opinion.faded }}
                 />
-                {hasContributor ? (
+                {hasContributor && (
                     <GridRow
                         left={contributor(contributorLeft)}
                         right={contributor(contributorRight)}
                         leftStyles={{
-                            backgroundColor: palette.neutral[100],
-                            verticalAlign: "bottom",
-                            borderLeft: `1px solid ${palette.opinion.main}`,
-                            borderBottom: `1px solid ${palette.opinion.main}`
+                            backgroundColor: palette.opinion.faded,
+                            verticalAlign: "bottom"
                         }}
                         rightStyles={{
-                            backgroundColor: palette.neutral[100],
-                            verticalAlign: "bottom",
-                            borderLeft: `1px solid ${palette.opinion.main}`,
-                            borderBottom: `1px solid ${palette.opinion.main}`
-                        }}
-                    />
-                ) : (
-                    <GridRow
-                        left="&nbsp;"
-                        right="&nbsp;"
-                        leftStyles={{
-                            backgroundColor: palette.neutral[100],
-                            verticalAlign: "bottom",
-                            borderLeft: `1px solid ${palette.opinion.main}`,
-                            borderBottom: `1px solid ${palette.opinion.main}`,
-                            lineHeight: "0"
-                        }}
-                        rightStyles={{
-                            backgroundColor: palette.neutral[100],
-                            verticalAlign: "bottom",
-                            borderLeft: `1px solid ${palette.opinion.main}`,
-                            borderBottom: `1px solid ${palette.opinion.main}`,
-                            lineHeight: "0"
+                            backgroundColor: palette.opinion.faded,
+                            verticalAlign: "bottom"
                         }}
                     />
                 )}
@@ -131,17 +96,17 @@ export const LinkGrid: React.FC<LinkGridProps> = ({ content, salt }) => {
     const rows = partition(content, 2).map((pair, i) => (
         <React.Fragment key={i}>
             <GridRow
-                left={<LinkCard content={pair[0]} />}
-                right={pair[1] ? <LinkCard content={pair[1]} /> : null}
+                left={<LinkCard content={pair[0]} theme="dark" />}
+                right={
+                    pair[1] ? (
+                        <LinkCard content={pair[1]} theme="light" />
+                    ) : null
+                }
                 leftStyles={{
-                    backgroundColor: palette.neutral[100],
-                    borderLeft: `1px solid ${palette.neutral[20]}`,
-                    borderBottom: `1px solid ${palette.neutral[20]}`
+                    backgroundColor: palette.neutral[20]
                 }}
                 rightStyles={{
-                    backgroundColor: palette.neutral[100],
-                    borderLeft: `1px solid ${palette.neutral[20]}`,
-                    borderBottom: `1px solid ${palette.neutral[20]}`
+                    backgroundColor: palette.neutral[86]
                 }}
             />
