+++ src/components/tests/opinion/collections/commentB/LinkCard.tsx
@@ -1,9 +1,10 @@
 import React from "react";
-import { Content } from "../../../api";
-import { TableRowCell } from "../../../layout/Table";
-import { TdCSS, LinkCSS, ImageCSS } from "../../../css";
-import { headline } from "../../../styles/typography";
-import { palette } from "@guardian/src-foundations";
+import { Content } from "../../../../../api";
+import { TableRowCell } from "../../../../../layout/Table";
+import { TdCSS, LinkCSS, ImageCSS } from "../../../../../css";
+import { headline } from "../../../../../styles/typography";
+
+type Theme = "light" | "dark";
 
 const brazeParameter = "?##braze_utm##"; // TODO solve link generation
 
@@ -11,12 +12,12 @@ const containerStyle: TdCSS = {
     padding: "10px 10px 30px 10px"
 };
 
-const fontStyle: LinkCSS = {
-    color: "#121212",
+const fontStyle = (theme: Theme): LinkCSS => ({
+    color: theme === "dark" ? "#dcdcdc" : "#121212",
     textDecoration: "none",
 
     ...headline({ level: 2 })
-};
+});
 
 const iconStyle: ImageCSS = {
     height: "1.4em",
@@ -26,18 +27,21 @@ const iconStyle: ImageCSS = {
 
 export const LinkCard: React.FC<{
     content: Content;
-}> = ({ content }) => {
+    theme: Theme;
+}> = ({ content, theme }) => {
     const webURL =
         "https://www.theguardian.com" +
         content.properties.href +
         brazeParameter; // TODO type curated content separately?
 
     const arrow =
-        "https://cdn.braze.eu/appboy/communication/assets/image_assets/images/5dcebdcb9ae1683cc77465a5/original.png?1573830091";
+        theme === "light"
+            ? "https://cdn.braze.eu/appboy/communication/assets/image_assets/images/5dcebdcb9ae1683cc77465a5/original.png?1573830091"
+            : "https://cdn.braze.eu/appboy/communication/assets/image_assets/images/5dcebdcb36dc7870b59d36b8/original.png?1573830091";
 
     return (
         <TableRowCell tdStyle={containerStyle}>
-            <a style={fontStyle} href={webURL}>
+            <a style={fontStyle(theme)} href={webURL}>
                 {content.header.headline}
                 <br />
                 <br />
