+++ src/components/tests/opinion/collections/commentC/Card.tsx
@@ -1,11 +1,16 @@
 import React from "react";
-import { FontCSS, TdCSS, ImageCSS } from "../../../css";
+import { FontCSS, TdCSS, ImageCSS } from "../../../../../css";
 import { palette } from "@guardian/src-foundations";
-import { Content, Tag } from "../../../api";
-import { formatImage } from "../../../image";
-import { kickerText } from "../../../kicker";
+import { Content, Tag } from "../../../../../api";
+import { formatImage } from "../../../../../image";
+import { kickerText } from "../../../../../kicker";
 import sanitizeHtml from "sanitize-html";
-import { Table, RowCell, TableRowCell, TableRow } from "../../../layout/Table";
+import {
+    Table,
+    RowCell,
+    TableRowCell,
+    TableRow
+} from "../../../../../layout/Table";
 
 type Size = "small" | "large";
 
@@ -41,11 +46,11 @@ const imgProfileStyle: ImageCSS = {
     border: "0"
 };
 
-const tdStyle: TdCSS = {
-    backgroundColor: palette.opinion.faded,
-    borderTop: `2px solid ${palette.opinion.main}`,
-    padding: "0"
-};
+const tdStyle = (isLarge: boolean): TdCSS => ({
+    padding: "0",
+    borderLeft: isLarge ? `1px solid ${palette.opinion.main}` : "none",
+    borderBottom: isLarge ? `1px solid ${palette.opinion.main}` : "none"
+});
 
 const metaWrapperStyle = (size: Size): TdCSS => {
     const rightPad = size === "large" ? "40px" : "10px";
@@ -211,10 +216,10 @@ const SupplementaryMeta: React.FC<{
     } else if (contributorImageSrc) {
         return (
             <RowCell>
-                <Table>
+                <TableRow>
                     <td style={{ width: "50%" }}></td>
                     {contributorImage}
-                </Table>
+                </TableRow>
             </RowCell>
         );
     }
@@ -287,7 +292,7 @@ export const Card: React.FC<Props> = ({
     const formattedImage = formatImage(
         image.url,
         salt,
-        size === "large" ? 600 : 300,
+        size === "large" ? 599 : 300,
         content.card.starRating
     );
 
@@ -315,7 +320,7 @@ export const Card: React.FC<Props> = ({
     });
 
     return (
-        <TableRowCell tdStyle={tdStyle}>
+        <TableRowCell tdStyle={tdStyle(size === "large")}>
             <Table>
                 {shouldShowImage && (
                     <RowCell tdStyle={{ padding: "0" }}>
@@ -323,7 +328,7 @@ export const Card: React.FC<Props> = ({
                             src={imageURL}
                             linkURL={webURL}
                             alt={imageAlt}
-                            width={size === "large" ? 600 : 294}
+                            width={size === "large" ? 599 : 294}
                         />
                     </RowCell>
                 )}
@@ -347,7 +352,7 @@ export const Card: React.FC<Props> = ({
                             contributor && contributor.properties.webTitle
                         }
                         size={size}
-                        width={size === "large" ? 180 : 147}
+                        width={size === "large" ? 179 : 146}
                     />
                 )}
             </Table>
@@ -371,7 +376,7 @@ export const ContributorImageWrapper: React.FC<{
     return (
         <ContributorImage
             salt={salt}
-            width={147}
+            width={146}
             src={profilePic}
             alt={contributor.properties.webTitle}
         />
