+++ src/components/tests/opinion/collections/commentC/LinkCard.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import { Content } from "../../../api";
-import { TableRowCell } from "../../../layout/Table";
-import { TdCSS, LinkCSS, ImageCSS } from "../../../css";
-import { headline } from "../../../styles/typography";
-
-type Theme = "light" | "dark";
+import { Content } from "../../../../../api";
+import { TableRowCell } from "../../../../../layout/Table";
+import { TdCSS, LinkCSS, ImageCSS } from "../../../../../css";
+import { headline } from "../../../../../styles/typography";
+import { palette } from "@guardian/src-foundations";
 
 const brazeParameter = "?##braze_utm##"; // TODO solve link generation
 
@@ -12,12 +11,12 @@ const containerStyle: TdCSS = {
     padding: "10px 10px 30px 10px"
 };
 
-const fontStyle = (theme: Theme): LinkCSS => ({
-    color: theme === "dark" ? "#dcdcdc" : "#121212",
+const fontStyle: LinkCSS = {
+    color: "#121212",
     textDecoration: "none",
 
     ...headline({ level: 2 })
-});
+};
 
 const iconStyle: ImageCSS = {
     height: "1.4em",
@@ -27,21 +26,18 @@ const iconStyle: ImageCSS = {
 
 export const LinkCard: React.FC<{
     content: Content;
-    theme: Theme;
-}> = ({ content, theme }) => {
+}> = ({ content }) => {
     const webURL =
         "https://www.theguardian.com" +
         content.properties.href +
         brazeParameter; // TODO type curated content separately?
 
     const arrow =
-        theme === "light"
-            ? "https://cdn.braze.eu/appboy/communication/assets/image_assets/images/5dcebdcb9ae1683cc77465a5/original.png?1573830091"
-            : "https://cdn.braze.eu/appboy/communication/assets/image_assets/images/5dcebdcb36dc7870b59d36b8/original.png?1573830091";
+        "https://cdn.braze.eu/appboy/communication/assets/image_assets/images/5dcebdcb9ae1683cc77465a5/original.png?1573830091";
 
     return (
         <TableRowCell tdStyle={containerStyle}>
-            <a style={fontStyle(theme)} href={webURL}>
+            <a style={fontStyle} href={webURL}>
                 {content.header.headline}
                 <br />
                 <br />
